change core asset max share supple to 10 billion change asset precision digits to 9

+++ libraries/chain/include/graphene/chain/protocol/confidential.hpp
@@ -149,7 +149,7 @@ struct transfer_to_blind_operation : public base_operation
 {
    struct fee_parameters_type { 
       uint64_t fee              = 5*GRAPHENE_BLOCKCHAIN_PRECISION; ///< the cost to register the cheapest non-free account
-      uint32_t price_per_output = 5*GRAPHENE_BLOCKCHAIN_PRECISION;
+      uint64_t price_per_output = 5*GRAPHENE_BLOCKCHAIN_PRECISION;
    };
 
 
@@ -237,7 +237,7 @@ struct blind_transfer_operation : public base_operation
 {
    struct fee_parameters_type { 
       uint64_t fee              = 5*GRAPHENE_BLOCKCHAIN_PRECISION; ///< the cost to register the cheapest non-free account
-      uint32_t price_per_output = 5*GRAPHENE_BLOCKCHAIN_PRECISION;
+      uint64_t price_per_output = 5*GRAPHENE_BLOCKCHAIN_PRECISION;
    };
 
    asset                 fee;
