+++ libraries/chain/include/graphene/chain/protocol/transfer.hpp
@@ -45,7 +45,7 @@ namespace graphene { namespace chain {
    {
       struct fee_parameters_type {
          uint64_t fee       = 20 * GRAPHENE_BLOCKCHAIN_PRECISION;
-         uint32_t price_per_kbyte = 10 * GRAPHENE_BLOCKCHAIN_PRECISION; /// only required for large memos.
+         uint64_t price_per_kbyte = 10 * GRAPHENE_BLOCKCHAIN_PRECISION; /// only required for large memos.
       };
 
       asset            fee;
